requirements.txt
asyncpg
aiogram==2.7
aiohttp
aioredis==1.3.1
python-dotenv~=1.0.0
pillow
Django~=4.2.3
psycopg2-binary
psycopg2
coverage
django-rest-framework
serializers
djangorestframework~=3.14.0
django-cors-headers